requirements.txt
alabaster==0.7.12
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.8.6
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
crispy-bootstrap5==0.6
cryptography==36.0.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.9
django-allauth==0.46.0
django-ckeditor==6.2.0
django-coverage-plugin==2.0.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.5
django-js-asset==1.2.2
django-model-utils==4.2.0
django-redis==5.0.0
django-stubs==1.8.0
django-stubs-ext==0.3.1
djangorestframework==3.12.4
djongo==1.3.6
docutils==0.17.1
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==9.8.3
filelock==3.4.0
flake8==4.0.1
flake8-isort==4.1.1
hiredis==2.0.0
identify==2.4.0
idna==3.3
imagesize==1.3.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.29.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==21.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymongo==3.12.1
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.2.4
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
